# Conflicts:
#	back/munggle/src/main/java/com/munggle/domain/exception/ExceptionMessage.java

+++ back/munggle/src/main/java/com/munggle/dog/service/DogServiceImpl.java
@@ -5,6 +5,9 @@ import com.munggle.dog.dto.DogDetailDto;
 import com.munggle.dog.dto.DogUpdateDto;
 import com.munggle.dog.mapper.DogMapper;
 import com.munggle.dog.repository.DogRepository;
+import com.munggle.domain.exception.DogNotFoundException;
+import com.munggle.domain.exception.ExceptionMessage;
+import com.munggle.domain.exception.NotYourDogException;
 import com.munggle.domain.exception.UserNotFoundException;
 import com.munggle.domain.model.entity.Dog;
 import com.munggle.domain.model.entity.User;
@@ -31,16 +34,18 @@ public class DogServiceImpl implements DogService {
     private final DogRepository dogRepository;
     private final UserRepository userRepository;
 
-    // 전달된 MultipartFile 이미지 업로드
+    // 전달된 MultipartFile 이미지 업로드 경로 : dog/반려견ID/
+
     String dogFilePath = "dog/";
 
-    public FileInfoDto dogImageUpload(Long dogId, MultipartFile file){
+    public FileInfoDto uploadDogImage(Long dogId, MultipartFile file){
         // 이미지 정보 저장
         String uploadPath = dogFilePath + dogId + "/";
         FileInfoDto fileInfoDto = fileS3UploadService.uploadFile(uploadPath, file);
 
         return fileInfoDto;
     }
+
     @Override
     @Transactional
     public void insertDog(DogCreateDto dogCreateDto) {
@@ -55,49 +60,65 @@ public class DogServiceImpl implements DogService {
         Long dogId = dogRepository.save(dog).getDogId();
 
         // 전달된 이미지가 있는 경우에만 수정
-        if(dogCreateDto.getImage() != null) {
+        if(dogCreateDto.getImage() != null && !dogCreateDto.getImage().isEmpty()) {
 
-            dog.updateImage(dogImageUpload(dogId, dogCreateDto.getImage()));
+            dog.updateImage(uploadDogImage(dogId, dogCreateDto.getImage()));
         }
     }
 
     @Override
     @Transactional
-    public void updateDog(Long dogId, DogUpdateDto dogUpdateDto) {
+    public void updateDog(Long userId, Long dogId, DogUpdateDto dogUpdateDto) {
         Dog dog = dogRepository.findByDogIdAndIsDeletedIsFalse(dogId)
-                .orElseThrow(()->new NoSuchElementException());
+                .orElseThrow(()->new DogNotFoundException(ExceptionMessage.DOG_NOT_FOUND));
         dog.updateDog(dogUpdateDto);
 
+        // 로그인 사용자의 반려견이 아닌 경우 예외처리
+        if(userId != dog.getUser().getId()){
+            throw new NotYourDogException(ExceptionMessage.NOT_YOUR_DOG);
+        }
+
         // 전달된 파일이 있는 경우에 기존 이미지 삭제 후 변경
-        if(dogUpdateDto.getImage() != null){
+        if(dogUpdateDto.getImage() != null && !dogUpdateDto.getImage().isEmpty()) {
 
             // 기존 파일 S3에서 삭제
             String uploadPath = dogFilePath + dogId + "/";
             fileS3UploadService.removeFolderFiles(uploadPath);
 
-            dog.updateImage(dogImageUpload(dogId, dogUpdateDto.getImage()));
+            dog.updateImage(uploadDogImage(dogId, dogUpdateDto.getImage()));
         }
     }
 
     @Override
     @Transactional
-    public void deleteDog(Long dogId) {
+    public void deleteDog(Long userId, Long dogId) {
         Dog dog = dogRepository.findByDogIdAndIsDeletedIsFalse(dogId)
-                .orElseThrow(()->new NoSuchElementException());
+                .orElseThrow(()->new DogNotFoundException(ExceptionMessage.DOG_NOT_FOUND));
+
+        // 로그인 사용자의 반려견이 아닌 경우 예외처리
+        if(userId != dog.getUser().getId()){
+            throw new NotYourDogException(ExceptionMessage.NOT_YOUR_DOG);
+        }
+
         dog.deleteDog();
     }
 
     @Override
     public DogDetailDto getDetailDog(Long dogId) {
         Dog dog = dogRepository.findByDogIdAndIsDeletedIsFalse(dogId)
-                .orElseThrow(()->new NoSuchElementException());
+                .orElseThrow(()->new DogNotFoundException(ExceptionMessage.DOG_NOT_FOUND));
         return DogMapper.toDetailDto(dog);
     }
 
     @Override
     public List<DogDetailDto> getDogList(Long userId) {
+
+        // 넘겨 받은 사용자가 있는지 확인
+        User user = userRepository.findByIdAndIsEnabledTrue(userId)
+                .orElseThrow(() -> new UserNotFoundException(USER_NOT_FOUND));
+        
         List<Dog> list = dogRepository.findAllByUserIdAndIsDeletedIsFalse(userId)
-                .orElseThrow(()->new NoSuchElementException());
+                .orElseThrow(()->new DogNotFoundException(ExceptionMessage.DOG_NOT_FOUND));
         return list.stream().map(dog -> DogMapper.toDetailDto(dog)).collect(Collectors.toList());
     }
 }
